UBR-255: Fix for error shown when sessions times out
When entering a passholder's identification number for the first time,
without having him in the passholder's cache and without a valid session
(session timed out or cookie manually removed) an error-page was shown.
Instead the user should be redirected to the login page.

The correct fix however involves having the backend sending the correct
statuscode, 401 instead of 403.